requirements2.txt
anyio==3.4.0
argon2-cffi==21.1.0
attrs==21.2.0
autopep8==1.6.0
awsebcli==3.20.2
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
botocore==1.21.65
cachelib==0.4.1
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.3
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
future==0.16.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
json5==0.9.6
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.12.0
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysql==0.0.3
mysqlclient==2.1.0
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.6
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.5.9
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==3.0.22
pycodestyle==2.8.0
pycparser==2.2
Pygments==2.10.0
pyparsing==3.0.6
pypiwin32==223; platform_system == "Windows"
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==302; platform_system == "Windows"
pywinpty==1.1.6; platform_system == "Windows"
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
semantic-version==2.8.5
Send2Trash==1.8.0
six==1.14.0
sniffio==1.2.0
SQLAlchemy==1.4.27
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==2.0.2
yapf==0.31.0
